refactor(tests): Tier 3 batch 2 — split inventory, toolchain, and stack template tests
- inventory_test.go → mock_test.go, scan_test.go, util_test.go
- toolchain_test.go → version_test.go, detection_test.go
- template_test.go trimmed → tags_test.go extracted
